+++ src/mulit/suggestlist.html
@@ -49,10 +49,8 @@
 
       var multi = new Multilist({
         render: '#J_MultiList',
-        source: {
-          multipleSelect: true,
-          store: store
-        }
+        elCls:'bui-select-list',
+        store: store
       }).render();
     })
